pyproject.toml
[tool.poetry]
name = "talengo-data"
version = "0.1.0"
description = ""
authors = ["alexerhardt <alex.erhardt.es@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.9.0"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
simple-salesforce = "^1.12.6"
azure-storage-blob = "^12.22.0"
tenacity = "^9.0.0"
itables = "^2.1.5"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
